working bolt demo
ok so this has a setup where the hologram arrow corresponds to the bolt that the person is working on.  they can only move the bolt that the arrow is hovering over to ensure that they are following the guidance. this is mainly a proof of concept to show how we can use the hologram arrow. sad thing is, this demo took longer to make than the hologram arrow, but that's life   ¯\_(ツ)_/¯